fixing tests, adding large tests
