+++ src/tests.rs
@@ -7,4 +7,5 @@ mod nibble_path;
 mod node_type;
 mod restore;
 mod tree_cache;
+mod update_proof;
 mod vectors;
